+++ src/js/functions.js
@@ -1,0 +1,3 @@
+function removeChildren(parent) {
+    [...parent.children].reverse().forEach((child) => child.remove());
+}